+++ src/Pages/Films.tsx
@@ -1,24 +1,17 @@
 import axios from "axios";
 import React from "react";
+import FilmTable from "../components/GetFilms";
+import GetFilms from "../components/GetFilms";
 
-
-const [errorFlag, setErrorFlag] = React.useState(false)
-const [users, setUsers] = React.useState < Array < User >> ([])
-const [errorMessage, setErrorMessage] = React.useState("")
 
 const Films = () => {
-    return (<h1>Films</h1>)
+    return (
+        <div>
+            <h1>Films</h1>
+            <GetFilms />
+        </div>
+    )
 }
 
-const getFilms = () => {
-    axios.get('http://localhost:3000/api/users')
-        .then((response) => {
-        setErrorFlag(false)
-        setErrorMessage("")
-        setUsers(response.data)
-    }, (error) => {
-        setErrorFlag(true)
-        setErrorMessage(error.toString())
-    })
-}
+
 export default Films;
